pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11.0,<3.12"
openai = "^1.7.2"
fastapi = "^0.103.2"
pydantic = "^2.0.1"
uvicorn = "^0.23.2"
asyncio = "^3.4.3"
pytest-mock = "*"
python-dotenv = "^1.0.0"
jq = "^1.6.0"
loguru = "^0.7.2"
pyjwt = {extras = ["crypto"], version = "*"}
pypdf = "^4.0.1"
redis = "^5.0.1"
langsmith = "^0.1.31"
langchain-core = "^0.1.44"
langchain-community = "^0.0.34"
langchain = "^0.1.16"
orjson = "^3.9.13"
langgraph = "^0.0.38"
sse-starlette = "^2.0.0"
langchain-anthropic = "^0.1.8"
langchain-google-vertexai = "^1.0.1"
fireworks-ai = "^0.11.2"
cohere = ">=4.32,<5.0"
urllib3 = "1.26.18"
boto3 = "^1.34.105"
typing-extensions = ">=3.7.4"
langserve = "^0.0.41"
semantic-router = { extras = ["processing"], version = ">=0.0.34" }
tiktoken = "^0.6.0"

alembic = "^1.11.1"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.27" }
asyncpg = "^0.28.0"
structlog = "^23.1.0"
rich = "^13.5.2"
colorama = "^0.4.6"
pydantic-settings = "^2.0.2"
asgi-correlation-id = "^4.2.0"
python-multipart = "^0.0.9"

duckduckgo-search = "^4.4"
wikipedia = "^1.4.0"
arxiv = "^2.1.0"
xmltodict = "^0.13.0"
python-magic = "^0.4.27"
beautifulsoup4 = "^4.12.3"
pdfminer-six = "^20221105"
psycopg2-binary = "^2.9.9"
qdrant-client = "^1.8.0"
azure-monitor-opentelemetry = "^1.3.0"
dataclasses-json = "^0.6.4"
unstructured-client = "^0.22.0"
langchain-robocorp = "^0.0.5"
langchain-openai = "^0.1.3"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
ruff = "^0.0.280"
pytest = "^7.4.0"
factory-boy = "^3.3.0"
httpx = "^0.25.2"
pytest-asyncio = "^0.21.1"
pytest-mock = "^3.11.1"
pytest-socket = "^0.6.0"
pytest-watch = "^4.2.0"
pytest-timeout = "^2.2.0"
pytest-cov = "^4.0.0"

